requirements.txt
Pillow>=10.0.0
piexif>=1.1.3
pytest>=7.4.0
tkinterdnd2>=0.4.2


